requirements.txt
azure-core==1.26.1
azure-identity==1.12.0
azure-search-documents==11.3.0
matplotlib==3.6.2
openai==0.25.0
pandas==1.5.1
plotly==5.11.0
python-dotenv==0.21.0
scikit-learn==1.1.3
scipy==1.9.3
streamlit==1.20.0
streamlit_chat==0.0.2.2
termcolor==2.2.0